+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/KafkaProxy.java
@@ -42,12 +42,12 @@ import io.kroxylicious.proxy.config.Configuration;
 import io.kroxylicious.proxy.config.IllegalConfigurationException;
 import io.kroxylicious.proxy.config.MicrometerDefinition;
 import io.kroxylicious.proxy.config.PluginFactoryRegistry;
-import io.kroxylicious.proxy.config.admin.AdminHttpConfiguration;
+import io.kroxylicious.proxy.config.admin.ManagementConfiguration;
 import io.kroxylicious.proxy.internal.ApiVersionsServiceImpl;
 import io.kroxylicious.proxy.internal.KafkaProxyInitializer;
 import io.kroxylicious.proxy.internal.MeterRegistries;
 import io.kroxylicious.proxy.internal.PortConflictDetector;
-import io.kroxylicious.proxy.internal.admin.AdminHttpInitializer;
+import io.kroxylicious.proxy.internal.admin.ManagementInitializer;
 import io.kroxylicious.proxy.internal.config.Features;
 import io.kroxylicious.proxy.internal.net.DefaultNetworkBindingOperationProcessor;
 import io.kroxylicious.proxy.internal.net.EndpointRegistry;
@@ -75,7 +75,7 @@ public final class KafkaProxy implements AutoCloseable {
     }
 
     private final @NonNull Configuration config;
-    private final @Nullable AdminHttpConfiguration adminHttpConfig;
+    private final @Nullable ManagementConfiguration managementConfiguration;
     private final @NonNull List<MicrometerDefinition> micrometerConfig;
     private final @NonNull List<VirtualClusterModel> virtualClusterModels;
     private final AtomicBoolean running = new AtomicBoolean();
@@ -85,14 +85,14 @@ public final class KafkaProxy implements AutoCloseable {
     private final @NonNull PluginFactoryRegistry pfr;
     private @Nullable MeterRegistries meterRegistries;
     private @Nullable FilterChainFactory filterChainFactory;
-    private @Nullable EventGroupConfig adminEventGroup;
+    private @Nullable EventGroupConfig managementEventGroup;
     private @Nullable EventGroupConfig serverEventGroup;
 
     public KafkaProxy(@NonNull PluginFactoryRegistry pfr, @NonNull Configuration config, @NonNull Features features) {
         this.pfr = requireNonNull(pfr);
         this.config = validate(requireNonNull(config), requireNonNull(features));
         this.virtualClusterModels = config.virtualClusterModel(pfr);
-        this.adminHttpConfig = config.adminHttpConfig();
+        this.managementConfiguration = config.management();
         this.micrometerConfig = config.getMicrometer();
     }
 
@@ -120,16 +120,17 @@ public final class KafkaProxy implements AutoCloseable {
             STARTUP_SHUTDOWN_LOGGER.info("Kroxylicious is starting");
 
             var portConflictDefector = new PortConflictDetector();
-            Optional<HostPort> adminHttpHostPort = Optional.ofNullable(adminHttpConfig != null ? new HostPort(adminHttpConfig.host(), adminHttpConfig.port()) : null);
-            portConflictDefector.validate(virtualClusterModels, adminHttpHostPort);
+            var managementHostPort = Optional.ofNullable(managementConfiguration)
+                    .map(c -> new HostPort(c.getEffectiveBindAddress(), c.getEffectivePort()));
+            portConflictDefector.validate(virtualClusterModels, managementHostPort);
 
             var availableCores = Runtime.getRuntime().availableProcessors();
             meterRegistries = new MeterRegistries(pfr, micrometerConfig);
 
-            this.adminEventGroup = buildNettyEventGroups("admin", availableCores, config.isUseIoUring());
+            this.managementEventGroup = buildNettyEventGroups("management", availableCores, config.isUseIoUring());
             this.serverEventGroup = buildNettyEventGroups("server", availableCores, config.isUseIoUring());
 
-            var managementFuture = maybeStartManagementListener(adminEventGroup, meterRegistries);
+            var managementFuture = maybeStartManagementListener(managementEventGroup, meterRegistries);
 
             var overrideMap = getApiKeyMaxVersionOverride(config);
             ApiVersionsServiceImpl apiVersionsService = new ApiVersionsServiceImpl(overrideMap);
@@ -217,16 +218,16 @@ public final class KafkaProxy implements AutoCloseable {
 
     @NonNull
     private CompletableFuture<Void> maybeStartManagementListener(EventGroupConfig eventGroupConfig, MeterRegistries meterRegistries) {
-        return Optional.ofNullable(adminHttpConfig)
-                .map(c -> {
+        return Optional.ofNullable(managementConfiguration)
+                .map(mc -> {
                     var metricsBootstrap = new ServerBootstrap().group(eventGroupConfig.bossGroup(), eventGroupConfig.workerGroup())
                             .option(ChannelOption.SO_REUSEADDR, true)
                             .channel(eventGroupConfig.clazz())
-                            .childHandler(new AdminHttpInitializer(meterRegistries, c));
-                    LOGGER.info("Binding management endpoint: {}:{}", c.host(), c.port());
+                            .childHandler(new ManagementInitializer(meterRegistries, mc));
+                    LOGGER.info("Binding management endpoint: {}:{}", mc.getEffectiveBindAddress(), mc.getEffectivePort());
 
                     var future = new CompletableFuture<Void>();
-                    metricsBootstrap.bind(adminHttpConfig.host(), adminHttpConfig.port())
+                    metricsBootstrap.bind(managementConfiguration.getEffectiveBindAddress(), managementConfiguration.getEffectivePort())
                             .addListener((ChannelFutureListener) channelFuture -> ForkJoinPool.commonPool().execute(() -> {
                                 // we complete on a separate thread so that any chained work won't get run on the Netty thread.
                                 if (channelFuture.cause() != null) {
@@ -266,8 +267,8 @@ public final class KafkaProxy implements AutoCloseable {
                 if (serverEventGroup != null) {
                     closeFutures.addAll(serverEventGroup.shutdownGracefully());
                 }
-                if (adminEventGroup != null) {
-                    closeFutures.addAll(adminEventGroup.shutdownGracefully());
+                if (managementEventGroup != null) {
+                    closeFutures.addAll(managementEventGroup.shutdownGracefully());
                 }
                 closeFutures.forEach(Future::syncUninterruptibly);
                 if (filterChainFactory != null) {
@@ -288,7 +289,7 @@ public final class KafkaProxy implements AutoCloseable {
             }
         }
         finally {
-            adminEventGroup = null;
+            managementEventGroup = null;
             serverEventGroup = null;
             meterRegistries = null;
             filterChainFactory = null;